refactor file system again
