+++ src/components/LanguageMenu/LanguageMenu.jsx
@@ -1,9 +1,10 @@
-import { useOpen } from "../../../utils/hooks";
-import PullOut from "../../Reuse/PullOut/PullOut";
-import lang from "../../../assets/language.svg";
-import russia from "../../../assets/russia.svg";
-import uk from "../../../assets/uk.svg";
-import ua from "../../../assets/ua.svg";
+import { useOpen } from "../../utils/hooks";
+import PullOut from "../Reuse/PullOut/PullOut";
+
+import lang from "../../assets/language.svg";
+import russia from "../../assets/russia.svg";
+import uk from "../../assets/uk.svg";
+import ua from "../../assets/ua.svg";
 
 import "./LanguageMenu.css";
